修复: CaiRewardChest玩家快速堆叠刷物品
