+++ CaiRewardChest/Plugin.cs
@@ -1,9 +1,7 @@
-﻿using System.Diagnostics;
-using System.Text;
+﻿using On.OTAPI;
 using Terraria;
 using TerrariaApi.Server;
 using TShockAPI;
-using Utils = TShockAPI.Utils;
 
 namespace CaiRewardChest;
 
@@ -28,11 +26,21 @@ public class CaiRewardChest : TerrariaPlugin
     {
         Db.Init();
         GetDataHandlers.ChestOpen.Register(OnChestOpen);
+        Hooks.Chest.InvokeQuickStack += ChestOnInvokeQuickStack;
         Commands.ChatCommands.Add(new Command("CaiRewardChest.Admin", InitChest, "初始化奖励箱"));
         Commands.ChatCommands.Add(new Command("CaiRewardChest.Admin", ClearChest, "清空奖励箱"));
         Commands.ChatCommands.Add(new Command("CaiRewardChest.Admin", DeleteChest, "删除奖励箱"));
         Commands.ChatCommands.Add(new Command("CaiRewardChest.Admin", AddChest, "添加奖励箱"));
         Commands.ChatCommands.Add(new Command("CaiRewardChest.Admin", EditChest, "编辑奖励箱"));
+    }
+
+    private bool ChestOnInvokeQuickStack(Hooks.Chest.orig_InvokeQuickStack orig, int playerid, Item item,
+        int chestindex)
+    {
+        RewardChest? chest = Db.GetChestById(chestindex);
+        if (chest != null)
+            return false;
+        return true;
     }
 
     private void EditChest(CommandArgs args)
@@ -115,7 +123,7 @@ public class CaiRewardChest : TerrariaPlugin
                 return;
             }
 
-            GiveItem(chest, e);
+            Utils.GiveItem(chest, e);
         }
 
         catch (Exception ex)
@@ -124,46 +132,6 @@ public class CaiRewardChest : TerrariaPlugin
         }
     }
 
-    public static int InventorySlotAvailableCount(TSPlayer plr)
-    {
-        int num = 0;
-        if (plr.RealPlayer)
-            for (int index = 0; index < 50; ++index)
-                if (plr.TPlayer.inventory[index] == null || !plr.TPlayer.inventory[index].active ||
-                    plr.TPlayer.inventory[index].Name == "")
-                    ++num;
-        return num;
-    }
-
-    public void GiveItem(RewardChest chest, GetDataHandlers.ChestOpenEventArgs args)
-    {
-        int chestType = WorldGen.GetChestItemDrop(chest.X, chest.Y, Main.tile[chest.X, chest.Y].type);
-        if (InventorySlotAvailableCount(args.Player) >=
-            chest.Chest.item.Count(i => i != null && i.netID != 0 && i.stack != 0) + 1)
-        {
-            foreach (Item? i in chest.Chest.item) args.Player.GiveItem(i.netID, i.stack, i.prefix);
-            List<string> itemsReceived = chest.Chest.item
-                .Where(i => i != null && i.netID != 0 && i.stack != 0)
-                .Select(i => TShock.Utils.ItemTag(i)).ToList();
-
-
-            itemsReceived.Add(TShock.Utils.ItemTag(new Item()
-            {
-                netID = chestType,
-                stack = 1
-            }));
-            args.Player.GiveItem(chestType, 1, 0);
-            args.Player.SendSuccessMessage($"[i:{chestType}]你打开了一个奖励箱: " +
-                                           $"" + string.Join(", ", itemsReceived));
-            chest.HasOpenPlayer.Add(args.Player.Account.ID);
-            Db.UpdateChest(chest);
-        }
-        else
-        {
-            args.Player.SendWarningMessage($"[i:{chestType}]你的背包格子不够哦," +
-                                           $"还需要清空{chest.Chest.item.Count(i => i != null && i.netID != 0 && i.stack != 0) + 1 - InventorySlotAvailableCount(args.Player)}个格子!");
-        }
-    }
 
     private void ClearChest(CommandArgs args)
     {
@@ -188,8 +156,11 @@ public class CaiRewardChest : TerrariaPlugin
 
     protected override void Dispose(bool disposing)
     {
-        if (disposing) GetDataHandlers.ChestOpen.UnRegister(OnChestOpen);
-
+        if (disposing)
+        {
+            GetDataHandlers.ChestOpen.UnRegister(OnChestOpen);
+            Hooks.Chest.InvokeQuickStack -= ChestOnInvokeQuickStack;
+        }
         base.Dispose(disposing);
     }
 }
